feat(l10n): add intl localization support for Arabic and English
This commit introduces localization support for Arabic and English by adding ARB files and generating the necessary Dart files for internationalization. Additionally, the `.gitignore` file is updated to include `.env`, and the `pubspec.yaml` is modified to include `flutter_gen` for better code generation. The `analysis_options.yaml` is also updated with additional linting rules for improved code quality.